refactor: share HttpApi parity helpers
